requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
bleach==6.1.0
click==8.1.7
Deprecated==1.2.14
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
limits==3.13.0
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.31.1
webencodings==0.5.1
wrapt==1.16.0
